weather-client/infra/requirements.txt
# AWS CDK Core
aws-cdk-lib==2.161.1
constructs>=10.0.0,<11.0.0

# CDK Nag for security best practices
cdk-nag==2.28.195

# Development dependencies
pytest==7.4.4
pytest-cov==4.1.0
